Move lacing converter logic to nodemodel. Remove x y prop serialization...for now.
