+++ src/Workspaces/Workspaces.cs
@@ -75,7 +75,6 @@ namespace Autodesk.Workspaces
                         new AnnotationConverter(),
                         new WorkspaceConverter(engineController, scheduler, factory, isTestMode, verboseLogging),
                         new NodeModelConverter(manager, libraryServices),
-                        new LacingStrategyConverter(),
                     },
                 ReferenceResolverProvider = () => { return new IdReferenceResolver(); }
             };
@@ -110,7 +109,6 @@ namespace Autodesk.Workspaces
                         new WorkspaceConverter(engineController, scheduler, factory,
                         isTestMode, verboseLogging),
                         new NodeModelConverter(manager, libraryServices),
-                        new LacingStrategyConverter(),
                     },
                 ReferenceResolverProvider = () => { return new IdReferenceResolver(); }
             };
